apps/orchestrator/requirements.txt
# Created automatically by Cursor AI (2025-08-25)
# FastAPI Orchestrator with CrewAI

# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# CrewAI and AI dependencies
crewai==0.11.0
langchain==0.1.0
langchain-openai==0.0.2
openai==1.3.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Redis and messaging
redis==5.0.1
nats-py==2.6.0

# AWS/S3
boto3==1.34.0
minio==7.2.0

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# GIS and spatial data
geopandas==0.14.1
shapely==2.0.2
pyproj==3.6.1
fiona==1.9.5

# Data processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
tenacity==8.2.3
pydantic==2.5.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
